requirements.txt
Flask
Flask-Session
python-dotenv
psycopg2-binary
requests
SQLAlchemy
